Move net-related types into net module
Move all configuration related to the network into the `net` module to ease discoverability and to keep things more organized.
